+++ src/main/java/com/amashchenko/maven/plugin/gitflow/AbstractGitFlowMojo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2018 Aleksandr Mashchenko.
+ * Copyright 2014-2019 Aleksandr Mashchenko.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -46,6 +46,8 @@ import org.codehaus.plexus.util.cli.Commandline;
 public abstract class AbstractGitFlowMojo extends AbstractMojo {
     /** A full name of the versions-maven-plugin set goal. */
     private static final String VERSIONS_MAVEN_PLUGIN_SET_GOAL = "org.codehaus.mojo:versions-maven-plugin:set";
+    /** A full name of the versions-maven-plugin set-property goal. */
+    private static final String VERSIONS_MAVEN_PLUGIN_SET_PROPERTY_GOAL = "org.codehaus.mojo:versions-maven-plugin:set-property";
     /** Name of the tycho-versions-plugin set-version goal. */
     private static final String TYCHO_VERSIONS_PLUGIN_SET_GOAL = "org.eclipse.tycho:tycho-versions-plugin:set-version";
 
@@ -134,6 +136,23 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
     private boolean versionsForceUpdate = false;
 
     /**
+     * Property to set version to.
+     *
+     * @since 1.12.1
+     */
+    @Parameter(property = "versionProperty")
+    private String versionProperty;
+
+    /**
+     * Whether to skip updating version. Useful with {@link #versionProperty} to be
+     * able to update <code>revision</code> property without modifying version tag.
+     * 
+     * @since 1.12.1
+     */
+    @Parameter(property = "skipUpdateVersion")
+    private boolean skipUpdateVersion = false;
+
+    /**
      * The path to the Maven executable. Defaults to "mvn".
      */
     @Parameter(property = "mvnExecutable")
@@ -146,7 +165,7 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
 
     /** Maven session. */
     @Parameter(defaultValue = "${session}", readonly = true)
-    private MavenSession mavenSession;
+    protected MavenSession mavenSession;
     /** Default prompter. */
     @Component
     protected Prompter prompter;
@@ -207,25 +226,36 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
      * @throws MojoFailureException
      */
     protected String getCurrentProjectVersion() throws MojoFailureException {
+        final Model model = readModel(mavenSession.getCurrentProject());
+        if (model.getVersion() == null) {
+            throw new MojoFailureException(
+                    "Cannot get current project version. This plugin should be executed from the parent project.");
+        }
+        return model.getVersion();
+    }
+
+    /**
+     * Reads model from Maven project pom.xml.
+     * 
+     * @param project
+     *            Maven project
+     * @return Maven model
+     * @throws MojoFailureException
+     */
+    private Model readModel(MavenProject project) throws MojoFailureException {
         try {
             // read pom.xml
-            final MavenXpp3Reader mavenReader = new MavenXpp3Reader();
-            final FileReader fileReader = new FileReader(mavenSession
-                    .getCurrentProject().getFile().getAbsoluteFile());
+            Model model;
+            FileReader fileReader = new FileReader(project.getFile().getAbsoluteFile());
+            MavenXpp3Reader mavenReader = new MavenXpp3Reader();
             try {
-                final Model model = mavenReader.read(fileReader);
-
-                if (model.getVersion() == null) {
-                    throw new MojoFailureException(
-                            "Cannot get current project version. This plugin should be executed from the parent project.");
-                }
-
-                return model.getVersion();
+                model = mavenReader.read(fileReader);
             } finally {
                 if (fileReader != null) {
                     fileReader.close();
                 }
             }
+            return model;
         } catch (Exception e) {
             throw new MojoFailureException("", e);
         }
@@ -265,16 +295,15 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
 
         List<MavenProject> projects = mavenSession.getProjects();
         for (MavenProject project : projects) {
-            builtArtifacts.add(project.getGroupId() + ":"
-                    + project.getArtifactId() + ":" + project.getVersion());
+            final Model model = readModel(project);
 
-            List<Dependency> dependencies = project.getDependencies();
+            builtArtifacts.add(model.getGroupId() + ":" + model.getArtifactId() + ":" + model.getVersion());
+
+            List<Dependency> dependencies = model.getDependencies();
             for (Dependency d : dependencies) {
-                String id = d.getGroupId() + ":" + d.getArtifactId() + ":"
-                        + d.getVersion();
-                if (!builtArtifacts.contains(id)
-                        && ArtifactUtils.isSnapshot(d.getVersion())) {
-                    snapshots.add(project + " -> " + d);
+                String id = d.getGroupId() + ":" + d.getArtifactId() + ":" + d.getVersion();
+                if (!builtArtifacts.contains(id) && ArtifactUtils.isSnapshot(d.getVersion())) {
+                    snapshots.add(model + " -> " + d);
                 }
             }
         }
@@ -564,6 +593,24 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
     }
 
     /**
+     * Replaces properties in message.
+     * 
+     * @param message
+     * @param map
+     *            Key is a string to replace wrapped in <code>@{...}</code>. Value
+     *            is a string to replace with.
+     * @return
+     */
+    private String replaceProperties(String message, Map<String, String> map) {
+        if (map != null) {
+            for (Entry<String, String> entr : map.entrySet()) {
+                message = StringUtils.replace(message, "@{" + entr.getKey() + "}", entr.getValue());
+            }
+        }
+        return message;
+    }
+
+    /**
      * Executes git commit -a -m.
      * 
      * @param message
@@ -582,20 +629,14 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
      * 
      * @param message
      *            Commit message.
-     * @param map
-     *            Key is a string to replace wrapped in <code>@{...}</code>.
-     *            Value is a string to replace with.
+     * @param messageProperties
+     *            Properties to replace in message.
      * @throws MojoFailureException
      * @throws CommandLineException
      */
-    protected void gitCommit(String message, Map<String, String> map)
+    protected void gitCommit(String message, Map<String, String> messageProperties)
             throws MojoFailureException, CommandLineException {
-        if (map != null) {
-            for (Entry<String, String> entr : map.entrySet()) {
-                message = StringUtils.replace(message, "@{" + entr.getKey()
-                        + "}", entr.getValue());
-            }
-        }
+        message = replaceProperties(message, messageProperties);
 
         if (gpgSignCommit) {
             getLog().info("Committing changes. GPG-signed.");
@@ -619,14 +660,25 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
      *            Merge with --no-ff.
      * @param ffonly
      *            Merge with --ff-only.
+     * @param message
+     *            Merge commit message.
+     * @param messageProperties
+     *            Properties to replace in message.
      * @throws MojoFailureException
      * @throws CommandLineException
      */
-    protected void gitMerge(final String branchName, boolean rebase, boolean noff, boolean ffonly)
+    protected void gitMerge(final String branchName, boolean rebase, boolean noff, boolean ffonly, String message,
+            Map<String, String> messageProperties)
             throws MojoFailureException, CommandLineException {
         String sign = "";
         if (gpgSignCommit) {
             sign = "-S";
+        }
+        String msgParam = "";
+        String msg = "";
+        if (StringUtils.isNotBlank(message)) {
+            msgParam = "-m";
+            msg = replaceProperties(message, messageProperties);
         }
         if (rebase) {
             getLog().info("Rebasing '" + branchName + "' branch.");
@@ -636,10 +688,10 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
             executeGitCommand("merge", "--ff-only", sign, branchName);
         } else if (noff) {
             getLog().info("Merging (--no-ff) '" + branchName + "' branch.");
-            executeGitCommand("merge", "--no-ff", sign, branchName);
+            executeGitCommand("merge", "--no-ff", sign, branchName, msgParam, msg);
         } else {
             getLog().info("Merging '" + branchName + "' branch.");
-            executeGitCommand("merge", sign, branchName);
+            executeGitCommand("merge", sign, branchName, msgParam, msg);
         }
     }
 
@@ -653,7 +705,7 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
      */
     protected void gitMergeNoff(final String branchName)
             throws MojoFailureException, CommandLineException {
-        gitMerge(branchName, false, true, false);
+        gitMerge(branchName, false, true, false, null, null);
     }
 
     /**
@@ -679,19 +731,14 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
      *            Tag message.
      * @param gpgSignTag
      *            Make a GPG-signed tag.
-     * @param map
-     *            Key is a string to replace wrapped in <code>@{...}</code>. Value
-     *            is a string to replace with.
+     * @param messageProperties
+     *            Properties to replace in message.
      * @throws MojoFailureException
      * @throws CommandLineException
      */
-    protected void gitTag(final String tagName, String message, boolean gpgSignTag, Map<String, String> map)
+    protected void gitTag(final String tagName, String message, boolean gpgSignTag, Map<String, String> messageProperties)
             throws MojoFailureException, CommandLineException {
-        if (map != null) {
-            for (Entry<String, String> entr : map.entrySet()) {
-                message = StringUtils.replace(message, "@{" + entr.getKey() + "}", entr.getValue());
-            }
-        }
+        message = replaceProperties(message, messageProperties);
 
         if (gpgSignTag) {
             getLog().info("Creating GPG-signed '" + tagName + "' tag.");
@@ -874,10 +921,10 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
      * @throws MojoFailureException
      * @throws CommandLineException
      */
-    protected void mvnSetVersions(final String version)
-            throws MojoFailureException, CommandLineException {
+    protected void mvnSetVersions(final String version) throws MojoFailureException, CommandLineException {
         getLog().info("Updating version(s) to '" + version + "'.");
 
+        String newVersion = "-DnewVersion=" + version;
         String g = "";
         String a = "";
         if (versionsForceUpdate) {
@@ -886,11 +933,24 @@ public abstract class AbstractGitFlowMojo extends AbstractMojo {
         }
 
         if (tychoBuild) {
-            executeMvnCommand(TYCHO_VERSIONS_PLUGIN_SET_GOAL, "-DnewVersion="
-                    + version, "-Dtycho.mode=maven");
+            String prop = "";
+            if (StringUtils.isNotBlank(versionProperty)) {
+                prop = "-Dproperties=" + versionProperty;
+                getLog().info("Updating property '" + versionProperty + "' to '" + version + "'.");
+            }
+
+            executeMvnCommand(TYCHO_VERSIONS_PLUGIN_SET_GOAL, prop, newVersion, "-Dtycho.mode=maven");
         } else {
-            executeMvnCommand(VERSIONS_MAVEN_PLUGIN_SET_GOAL, g, a, "-DnewVersion="
-                    + version, "-DgenerateBackupPoms=false");
+            if (!skipUpdateVersion) {
+                executeMvnCommand(VERSIONS_MAVEN_PLUGIN_SET_GOAL, g, a, newVersion, "-DgenerateBackupPoms=false");
+            }
+
+            if (StringUtils.isNotBlank(versionProperty)) {
+                getLog().info("Updating property '" + versionProperty + "' to '" + version + "'.");
+
+                executeMvnCommand(VERSIONS_MAVEN_PLUGIN_SET_PROPERTY_GOAL, newVersion, "-Dproperty=" + versionProperty,
+                        "-DgenerateBackupPoms=false");
+            }
         }
     }