+++ src/main/java/com/amashchenko/maven/plugin/gitflow/GitFlowReleaseFinishMojo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2018 Aleksandr Mashchenko.
+ * Copyright 2014-2019 Aleksandr Mashchenko.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,7 +28,7 @@ import org.codehaus.plexus.util.StringUtils;
 
 /**
  * The git flow release finish mojo.
- * 
+ *
  */
 @Mojo(name = "release-finish", aggregator = true)
 public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
@@ -43,7 +43,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
 
     /**
      * Whether to skip calling Maven test goal before merging the branch.
-     * 
+     *
      * @since 1.0.5
      */
     @Parameter(property = "skipTestProject", defaultValue = "false")
@@ -51,7 +51,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
 
     /**
      * Whether to allow SNAPSHOT versions in dependencies.
-     * 
+     *
      * @since 1.2.2
      */
     @Parameter(property = "allowSnapshots", defaultValue = "false")
@@ -60,7 +60,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
     /**
      * Whether to rebase branch or merge. If <code>true</code> then rebase will
      * be performed.
-     * 
+     *
      * @since 1.2.3
      */
     @Parameter(property = "releaseRebase", defaultValue = "false")
@@ -68,7 +68,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
 
     /**
      * Whether to use <code>--no-ff</code> option when merging.
-     * 
+     *
      * @since 1.2.3
      */
     @Parameter(property = "releaseMergeNoFF", defaultValue = "true")
@@ -76,7 +76,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
 
     /**
      * Whether to push to the remote.
-     * 
+     *
      * @since 1.3.0
      */
     @Parameter(property = "pushRemote", defaultValue = "true")
@@ -84,7 +84,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
 
     /**
      * Whether to use <code>--ff-only</code> option when merging.
-     * 
+     *
      * @since 1.4.0
      */
     @Parameter(property = "releaseMergeFFOnly", defaultValue = "false")
@@ -92,7 +92,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
 
     /**
      * Whether to remove qualifiers from the next development version.
-     * 
+     *
      * @since 1.6.0
      */
     @Parameter(property = "digitsOnlyDevVersion", defaultValue = "false")
@@ -101,7 +101,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
     /**
      * Development version to use instead of the default next development
      * version in non interactive mode.
-     * 
+     *
      * @since 1.6.0
      */
     @Parameter(property = "developmentVersion", defaultValue = "")
@@ -110,7 +110,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
     /**
      * Which digit to increment in the next development version. Starts from
      * zero.
-     * 
+     *
      * @since 1.6.0
      */
     @Parameter(property = "versionDigitToIncrement")
@@ -120,7 +120,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
      * Whether to commit development version when starting the release (vs when
      * finishing the release which is the default). Has effect only when there
      * are separate development and production branches.
-     * 
+     *
      * @since 1.7.0
      */
     @Parameter(property = "commitDevelopmentVersionAtStart", defaultValue = "false")
@@ -129,7 +129,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
     /**
      * Maven goals to execute in the release branch before merging into the
      * production branch.
-     * 
+     *
      * @since 1.8.0
      */
     @Parameter(property = "preReleaseGoals")
@@ -137,7 +137,7 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
 
     /**
      * Maven goals to execute in the production branch after the release.
-     * 
+     *
      * @since 1.8.0
      */
     @Parameter(property = "postReleaseGoals")
@@ -145,20 +145,20 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
 
     /**
      * Whether to make a GPG-signed tag.
-     * 
+     *
      * @since 1.9.0
      */
     @Parameter(property = "gpgSignTag", defaultValue = "false")
     private boolean gpgSignTag = false;
 
     /**
-     * Whether this is use snapshot in release.
-     * 
+     * Whether to use snapshot in release.
+     *
      * @since 1.10.0
      */
-    @Parameter(defaultValue = "false")
-    protected boolean useSnapshotInRelease;
-    
+    @Parameter(property = "useSnapshotInRelease", defaultValue = "false")
+    private boolean useSnapshotInRelease;
+
     /** {@inheritDoc} */
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
@@ -207,37 +207,39 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
                 }
             }
 
-            if (!skipTestProject) {
-                // git checkout release/...
-                gitCheckout(releaseBranch);
+            // git checkout release/...
+            gitCheckout(releaseBranch);
 
+            if (!skipTestProject) {
                 // mvn clean test
                 mvnCleanTest();
             }
 
             // maven goals before merge
             if (StringUtils.isNotBlank(preReleaseGoals)) {
-                gitCheckout(releaseBranch);
-
                 mvnRun(preReleaseGoals);
             }
 
             String currentReleaseVersion = getCurrentProjectVersion();
+
+            Map<String, String> messageProperties = new HashMap<String, String>();
+            messageProperties.put("version", currentReleaseVersion);
+
             if (useSnapshotInRelease && ArtifactUtils.isSnapshot(currentReleaseVersion)) {
                 String commitVersion = currentReleaseVersion.replace("-" + Artifact.SNAPSHOT_VERSION, "");
 
                 mvnSetVersions(commitVersion);
 
-                Map<String, String> properties = new HashMap<String, String>();
-                properties.put("version", commitVersion);
+                messageProperties.put("version", commitVersion);
 
-                gitCommit(commitMessages.getReleaseFinishMessage(), properties);
+                gitCommit(commitMessages.getReleaseFinishMessage(), messageProperties);
             }
 
             // git checkout master
             gitCheckout(gitFlowConfig.getProductionBranch());
 
-            gitMerge(releaseBranch, releaseRebase, releaseMergeNoFF, releaseMergeFFOnly);
+            gitMerge(releaseBranch, releaseRebase, releaseMergeNoFF, releaseMergeFFOnly,
+                    commitMessages.getReleaseFinishMergeMessage(), messageProperties);
 
             // get current project version from pom
             final String currentVersion = getCurrentProjectVersion();
@@ -249,12 +251,11 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
                             + Artifact.SNAPSHOT_VERSION, "");
                 }
 
-                Map<String, String> properties = new HashMap<String, String>();
-                properties.put("version", tagVersion);
+                messageProperties.put("version", tagVersion);
 
                 // git tag -a ...
                 gitTag(gitFlowConfig.getVersionTagPrefix() + tagVersion,
-                        commitMessages.getTagReleaseMessage(), gpgSignTag, properties);
+                        commitMessages.getTagReleaseMessage(), gpgSignTag, messageProperties);
             }
 
             // maven goals after merge
@@ -266,7 +267,27 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
                 // git checkout develop
                 gitCheckout(gitFlowConfig.getDevelopmentBranch());
 
-                gitMerge(releaseBranch, releaseRebase, releaseMergeNoFF, false);
+                // get develop version
+                final String developReleaseVersion = getCurrentProjectVersion();
+                if (commitDevelopmentVersionAtStart && useSnapshotInRelease) {
+                    // updating develop poms to master version to avoid merge conflicts
+                    mvnSetVersions(currentVersion);
+
+                    // commit the changes
+                    gitCommit(commitMessages.getUpdateDevToAvoidConflictsMessage());
+                }
+
+                // merge branch master into develop
+                gitMerge(releaseBranch, releaseRebase, releaseMergeNoFF, false,
+                        commitMessages.getReleaseFinishDevMergeMessage(), messageProperties);
+
+                if (commitDevelopmentVersionAtStart && useSnapshotInRelease) {
+                    // updating develop poms version back to pre merge state
+                    mvnSetVersions(developReleaseVersion);
+
+                    // commit the changes
+                    gitCommit(commitMessages.getUpdateDevBackPreMergeStateMessage());
+                }
             }
 
             if (commitDevelopmentVersionAtStart && !notSameProdDevName()) {
@@ -300,11 +321,10 @@ public class GitFlowReleaseFinishMojo extends AbstractGitFlowMojo {
                 // mvn versions:set -DnewVersion=... -DgenerateBackupPoms=false
                 mvnSetVersions(nextSnapshotVersion);
 
-                Map<String, String> properties = new HashMap<String, String>();
-                properties.put("version", nextSnapshotVersion);
+                messageProperties.put("version", nextSnapshotVersion);
 
                 // git commit -a -m updating for next development version
-                gitCommit(commitMessages.getReleaseFinishMessage(), properties);
+                gitCommit(commitMessages.getReleaseFinishMessage(), messageProperties);
             }
 
             if (installProject) {